+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 1.0.8
+
+* open_file package remove
+
+## 1.0.7
+
+* open_file package remove
+
+## 1.0.6
+
+* Updated dependencies
+
 ## 1.0.5
 
 * Updated dependencies, Bug fixes